Moving SyncOfstream to a separate file.
