+++ generator/osm_source.cpp
@@ -2,6 +2,7 @@
 #include "generator/feature_generator.hpp"
 #include "generator/intermediate_data.hpp"
 #include "generator/intermediate_elements.hpp"
+#include "generator/sync_ofsteam.hpp"
 #include "generator/osm_element.hpp"
 #include "generator/osm_o5m_source.hpp"
 #include "generator/osm_source.hpp"
@@ -532,30 +533,6 @@ private:
   }
 };
 }  // anonymous namespace
-
-void SyncOfstream::Open(string const & fullPath)
-{
-  lock_guard<mutex> guard(m_mutex);
-  m_stream.open(fullPath, std::ofstream::out);
-}
-
-bool SyncOfstream::IsOpened()
-{
-  lock_guard<mutex> guard(m_mutex);
-  return m_stream.is_open() && !m_stream.fail();
-}
-
-void SyncOfstream::Write(uint32_t featureId, vector<osm::Id> const & osmIds)
-{
-  if (!IsOpened())
-    return;
-
-  lock_guard<mutex> guard(m_mutex);
-  m_stream << featureId;
-  for (osm::Id const & osmId : osmIds)
-    m_stream << "," << osmId.OsmId();
-  m_stream << endl;
-}
 
 unique_ptr<EmitterBase> MakeMainFeatureEmitter(feature::GenerateInfo const & info)
 {
