+++ generator/osm_source.hpp
@@ -2,13 +2,11 @@
 
 #include "generator/generate_info.hpp"
 #include "generator/osm_element.hpp"
-#include "generator/osm_id.hpp"
 
-#include "std/fstream.hpp"
 #include "std/function.hpp"
 #include "std/iostream.hpp"
-#include "std/mutex.hpp"
 #include "std/unique_ptr.hpp"
+#include "std/vector.hpp"
 
 class SourceReader
 {
@@ -49,17 +47,6 @@ public:
   /// Sets buckets (mwm names).
   // TODO(syershov): Make this topic clear.
   virtual void GetNames(vector<string> & names) const = 0;
-};
-
-class SyncOfstream
-{
-  ofstream m_stream;
-  mutex m_mutex;
-
-public:
-  void Open(string const & fullPath);
-  bool IsOpened();
-  void Write(uint32_t featureId, vector<osm::Id> const & osmIds);
 };
 
 unique_ptr<EmitterBase> MakeMainFeatureEmitter(feature::GenerateInfo const & info);
